Minimum Requirements Met!
Scripts:
- Modified BoardManager:
This script now manages tasks such as creating the grid, filling in the grid with tiles, and deleting them.
- Modified Tile
This script now checks for adjacent tiles to the tile this script is attached to, and handles mouse events when we click on a tile.
- Added ClearableTile:
A component that is attached to a tile that we wish to delete.
- Added MovableTile:
A component that is attached to a tile to make it movable.
- Added TileColor
A component which gives a type of colour to the tile it is attached to.

Animations:
- Added an animation clip for when a tile is deleted.